+++ fpdf/encryption.py
@@ -11,6 +11,8 @@ import logging
 import math
 import stringprep
 import unicodedata
+from binascii import hexlify
+from codecs import BOM_UTF16_BE
 from os import urandom
 from typing import Callable, Iterable, Type, Union
 
@@ -220,7 +222,11 @@ class StandardSecurityHandler:
         if self.encryption_method == EncryptionMethod.NO_ENCRYPTION:
             return PDFString(string, encrypt=False).serialize()
         LOGGER.debug("Encrypting string: %s", string)
-        return f"<{bytes(self.encrypt_bytes(string.encode('latin-1'), obj_id)).hex().upper()}>"
+        try:
+            string.encode("latin-1")
+            return f"<{bytes(self.encrypt_bytes(string.encode('latin-1'), obj_id)).hex().upper()}>"
+        except UnicodeEncodeError:
+            return f'<{hexlify(bytearray(self.encrypt_bytes(BOM_UTF16_BE + string.encode("utf-16-be"), obj_id))).decode("latin-1")}>'
 
     def encrypt_stream(self, stream: bytes, obj_id: int) -> bytes:
         if self.encryption_method == EncryptionMethod.NO_ENCRYPTION: